pyproject.toml
[tool.poetry]
name = "kampan"
version = "0.0.1"
description = ""
authors = ["R202"]

[tool.poetry.dependencies]
python = ">=3.12"
Werkzeug = "^2.0.1"
Flask-Login = "^0.6.0"
watchdog = "^2.1.3"
Authlib = "^1.1.0"
numpy = "^1.22.3"
brython = "^3.10.7"
requests = "^2.28.1"
Markdown = "^3.4.1"
flask-me = "^1.0.2"
wtform = "^1.0"
flask-wtf = "^1.1.1"
lxml = "^5.1.0"
rq = "^1.15.1"
pandas = "^2.2.0"
openpyxl = "^3.1.2"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^3.9.2"
black = "^21.7b0"

[tool.poetry.scripts]
kampan-web = "kampan.cmd.web:main"
kampan-worker = "kampan.cmd.worker:main"
kampan-controller = "kampan.cmd.controller:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
